Added removing article image when deleting article

+++ server/utils/fs.js
@@ -1,4 +1,6 @@
 import fs from 'fs';
+import * as R from 'ramda';
+import { protectedFilePaths } from './constants';
 
 export const mkdirIfNotExistSync = path => {
   if (!fs.existsSync(path)) {
@@ -13,6 +15,10 @@ export const unlinkIfExistSync = filePath => {
 };
 
 export const unlinkIfExist = filePath => {
+  if (R.any(R.endsWith(R.__, filePath), protectedFilePaths)) {
+    return;
+  }
+
   fs.stat(filePath, err => {
     if (!err) {
       fs.unlink(filePath, () => {
